requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Deprecated==1.2.18
exceptiongroup==1.2.2
flake8==7.2.0
google-auth==2.39.0
google-genai==1.13.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyflakes==3.3.2
PyGithub==2.6.1
PyJWT==2.10.1
PyNaCl==1.5.0
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
wrapt==1.17.2
